--- src/PuzzGen.cpp
@@ -1,10 +1,0 @@
-// PuzzGen.cpp : Definiert die exportierten Funktionen für die DLL-Anwendung.
-//
-
-#include "PuzzGen.h"
-
-namespace PuzzGen 
-{
-
-}
-
