+++ src/PuzzGen.h
@@ -1,19 +1,19 @@
 // PuzzGenLibrary.h - Contains declaration of Function class  
 #pragma once  
 
-#include "Core.h"
+#include "core/Core.h"
+
+#include "core/EventManager.h"
+#include "core/Logger.h"
+#include "core/GraphNode.h"
+#include "core/Node.h"
+#include "core/Relation.h"
+#include "core/Randomizer.h"
 
 #include "Object.h"
-#include "Puzzle.h"
 #include "Rule.h"
-#include "Generator.h"
-#include "GraphNode.h"
-#include "Node.h"
-#include "Relation.h"
-#include "Randomizer.h"
 #include "Event.h"
-#include "EventManager.h"
-#include "Logger.h"
+
+#include "Puzzle.h"
 #include "Context.h"
-
-
+#include "Generator.h"
