+++ src/core/Core.h
