+++ src/core/GeneratorHelper.cpp
