+++ src/core/GraphNode.cpp
