+++ src/core/ObjectMetaData.h
